Refactor citation selection state into hook
